# Conflicts:
#	src/Arch/Mips/MipsDisassembler.cs
#	src/Arch/Mips/MipsProcessorArchitecture.cs
#	src/Arch/Mips/MipsRewriter.System.cs
#	src/UnitTests/Arch/Mips/MipsRewriterTests.cs
#	src/tools/reko.config
#	subjects/Elf-MIPS/redir.c
#	subjects/Elf-Sparc/rtems/sparc-rtems-unprotoize.h

+++ src/Decompiler/Analysis/ConditionCodeEliminator.cs
@@ -314,7 +314,7 @@ namespace Reko.Analysis
             var expNewHi = m.Slice(
                 PrimitiveType.CreateWord(tmpLo.DataType.Size),
                 sidTmp.Identifier,
-                (uint)tmpHi.DataType.BitSize);
+                tmpHi.DataType.BitSize);
             var stmNewHi = block.Statements.Insert(
                 iRolc + 3,
                 sidOrigHi.DefStatement.LinearAddress,
@@ -391,7 +391,7 @@ namespace Reko.Analysis
             var expNewHi = m.Slice(
                 PrimitiveType.CreateWord(tmpHi.DataType.Size),
                 sidTmp.Identifier,
-                (uint)tmpLo.DataType.BitSize);
+                tmpLo.DataType.BitSize);
             var stmNewHi = block.Statements.Insert(
                 iRorc + 2,
                 sidOrigHi.DefStatement.LinearAddress,
